requirements.txt
Flask
Flask-Login
Flask-PyMongo
python-dotenv
oauthlib
requests
mediapipe
opencv-python
pillow
numpy

